telegram_bot/requirements.txt
cryptography
python-dotenv
python-telegram-bot